convert user-groups fixture to encapsulated class
