+++ corehq/apps/users/models.py
@@ -38,7 +38,6 @@ from corehq.apps.users.util import (
     user_data_from_registration_form,
     user_display_string,
 )
-from corehq.apps.users.xml import group_fixture
 from corehq.apps.users.tasks import tag_docs_as_deleted, tag_forms_as_deleted_rebuild_associated_cases
 from corehq.apps.users.exceptions import InvalidLocationConfig
 from corehq.apps.sms.mixin import (
@@ -1665,29 +1664,6 @@ class CommCareUser(CouchUser, SingleMembershipMixin, CommCareMobileContactMixin)
             case.doc_type = chop_suffix(case.doc_type)
             case.save()
         self.save()
-
-    def get_group_fixture(self, last_sync=None):
-        def _should_sync_groups(groups, last_sync):
-            """
-            Determine if we need to sync the groups fixture by checking
-            the modified date on all groups compared to the
-            last sync.
-            """
-            if not last_sync or not last_sync.date:
-                return True
-
-            for group in groups:
-                if not group.last_modified or group.last_modified >= last_sync.date:
-                    return True
-
-            return False
-
-        groups = self.get_case_sharing_groups()
-
-        if _should_sync_groups(groups, last_sync):
-            return group_fixture(groups, self)
-        else:
-            return None
 
     def get_case_sharing_groups(self):
         from corehq.apps.groups.models import Group
